reorginized tests
